+++ tests/PropertyClass.php
@@ -1,18 +1,22 @@
 <?php
-class PropertyClass {
+class PropertyClass
+{
     private $_prop;
 
-    public function __construct($prop) {
+    public function __construct($prop)
+    {
         $this->_prop = $prop;
     }
-    public function __get($name) {
-        if($name == 'property') {
+    public function __get($name)
+    {
+        if($name == 'property')
+        {
             return $this->_prop;
         }
         return null;
     }
-    
-    public function __isset($name) {
+    public function __isset($name)
+    {
         return $name == 'property';
     }
 };